Changed the XML-Path to the Debug Settings. Moved the Observer to its own Observer Directory
