+++ etc/frontend/events.xml
@@ -8,8 +8,8 @@
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 -->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../lib/internal/Magento/Framework/Event/etc/events.xsd">
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
     <event name="controller_action_predispatch">
-        <observer name="magepsycho_easypathhints_controller_action_predispatch" instance="MagePsycho\Easypathhints\Model\Observer" method="controllerActionPredispatch" />
+        <observer name="magepsycho_easypathhints_controller_action_predispatch" instance="MagePsycho\Easypathhints\Observer\EasyTemplateHintsObserver" />
     </event>
 </config>
